Rename Retirement URL to plural
